Refactoring according to remarks of F.S.
Change the name of the test Test_simplePilotLogger.py.
Also, use unittest assert for test.
Also, call the test directly in ci_test.sh
Also, remove unecessary shell functions
:
